recipes-bsp: u-boot: add fw-utils v2018.1 tailored for Neutis
A hacky recipe that fetches code from master and uses the pyro's recipe
as a base. That's why we need to override do_compile.

In previous U-boot releases fw_printenv and el. were compiled with
make env. Now it's make envtools. So take it or leave it.
We'll drop this piece as soon as we move to the newest release.
Sumo should have everything we need.