support: reflash-utility: fix buffer overflow
When we send something that is bigger than 128 we trigger buffer
overflow. strlen(buff) might return 128 because there's a chance there's
'\0' after buff[]. Making it read less will fix the overflow.
The circular buffer is not ideal, though. I can come up with a valid
stream that won't trigger a reboot.