bsp: neutis-initramfs: fix resize bug
Resizing contains 2 steps:
1. parted resize 100
2. resize2fs /dev/blockdev

No need to execute these steps every boot, so there is a flag for skipping.
We check the size of the partition, we will skip resizing if it is 100%.
This check will pass if the parted was called successfully, but it
doesn't guarantee the true result of resize2fs call.
The size of file system might be not 100%, but the real partition size is 100%.

This patch fixes this issue. The resize2fs can determine the size of FS and
skip the fs-resizing request.