bsp: kernel: linux: 4.17: explicitly depend on bison-native
After migration to linux 4.17 we have a problem with a clean
build of linux-sunxi fail due to missing defconfig in the build dir.
It's a result of the failed execution of merge_config.sh as 4.16 stopped
of using pre-generated files and now requires bison to be available:
|   HOSTCC  scripts/basic/fixdep
|   GEN     ./Makefile
|   HOSTCC  scripts/kconfig/conf.o
|   YACC    scripts/kconfig/zconf.tab.c
| /bin/sh: bison: command not found

There is a "fix" in kernel.bbclass that adds bison-native to DEPENDS,
but anyway it doesn't get installed before do_kernel_configme task,
because bitbake installs required recipes only before do_configure.
https://patchwork.openembedded.org/patch/148171/